+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/transfer/TransferItem.java
@@ -1,6 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.core.data.impl.transfer;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 import javax.persistence.MappedSuperclass;
@@ -14,10 +15,11 @@ public abstract class TransferItem<T extends Boxable> implements Serializable {
   private static final long serialVersionUID = 1L;
 
   private Boolean received;
-
   private Boolean qcPassed;
-
   private String qcNote;
+  private BigDecimal distributedVolume;
+  private String distributedBoxAlias;
+  private String distributedBoxPosition;
 
   public abstract Transfer getTransfer();
 
@@ -47,13 +49,37 @@ public abstract class TransferItem<T extends Boxable> implements Serializable {
     this.qcNote = qcNote;
   }
 
+  public BigDecimal getDistributedVolume() {
+    return distributedVolume;
+  }
+
+  public void setDistributedVolume(BigDecimal distributedVolume) {
+    this.distributedVolume = distributedVolume;
+  }
+
+  public String getDistributedBoxAlias() {
+    return distributedBoxAlias;
+  }
+
+  public void setDistributedBoxAlias(String distributedBoxAlias) {
+    this.distributedBoxAlias = distributedBoxAlias;
+  }
+
+  public String getDistributedBoxPosition() {
+    return distributedBoxPosition;
+  }
+
+  public void setDistributedBoxPosition(String distributedBoxPosition) {
+    this.distributedBoxPosition = distributedBoxPosition;
+  }
+
   public abstract T getItem();
 
   public abstract void setItem(T item);
 
   @Override
   public int hashCode() {
-    return Objects.hash(received, qcPassed, qcNote, getItem());
+    return Objects.hash(received, qcPassed, qcNote, distributedVolume, distributedBoxAlias, distributedBoxPosition, getItem());
   }
 
   @Override
@@ -62,6 +88,9 @@ public abstract class TransferItem<T extends Boxable> implements Serializable {
         TransferItem::isReceived,
         TransferItem::isQcPassed,
         TransferItem::getQcNote,
+        TransferItem::getDistributedVolume,
+        TransferItem::getDistributedBoxAlias,
+        TransferItem::getDistributedBoxPosition,
         TransferItem::getItem);
   }